+++ src/controller_transaction.c
@@ -586,16 +586,17 @@ controller_transaction_wait(clixon_handle h,
     while ((dh = device_handle_each(h, dh)) != NULL){
         if (device_handle_tid_get(dh) != tid)
             continue;
-        if (device_handle_conn_state_get(dh) == CS_PUSH_CHECK ||
+        if (
+#ifdef CONTROLLER_PUSH_LOCK
+            device_handle_conn_state_get(dh) == CS_PUSH_LOCK ||
+#endif
+            device_handle_conn_state_get(dh) == CS_PUSH_CHECK ||
             device_handle_conn_state_get(dh) == CS_PUSH_EDIT ||
             device_handle_conn_state_get(dh) == CS_PUSH_VALIDATE)
             notready++;
-        if (device_handle_conn_state_get(dh) == CS_PUSH_WAIT)
+        else if (device_handle_conn_state_get(dh) == CS_PUSH_WAIT)
             wait++;
-        if (device_handle_conn_state_get(dh) != CS_PUSH_CHECK &&
-            device_handle_conn_state_get(dh) != CS_PUSH_EDIT &&
-            device_handle_conn_state_get(dh) != CS_PUSH_VALIDATE &&
-            device_handle_conn_state_get(dh) != CS_PUSH_WAIT)
+        else
             other++;
     }
     if ((notready||wait) && other){