+++ src/controller_transaction.h
@@ -66,16 +66,12 @@ typedef struct controller_transaction_t controller_transaction;
 
 /*! Transaction failed device close parameter
  *
- * If device handle is set, one can either:
- * - Ignore the device just fail transaction. Typically if device is already closed or handled
- *   elsewhere
- * - Device leaves the transaction
- * - Close the device and leave the transaction
+ * If device handle is set, one can ignore, leave or close
  */
 enum tr_failed_devclose_t {
-    TR_FAILED_DEV_IGNORE,
-    TR_FAILED_DEV_LEAVE,
-    TR_FAILED_DEV_CLOSE
+    TR_FAILED_DEV_IGNORE,  /* Ignore device just fail transaction, if device already closed */
+    TR_FAILED_DEV_LEAVE,   /* Device leaves the transaction */
+    TR_FAILED_DEV_CLOSE    /* Close the device and leave the transaction */
 };
 typedef enum tr_failed_devclose_t tr_failed_devclose;